refactor: Move <InlinedVideoList/> to @chainlit/components
