--- frontend/src/components/atoms/element/inlined/videos.tsx
@@ -1,17 +1,0 @@
-import { IVideoElement } from '@chainlit/components';
-
-import VideoElement from '../video';
-import { ListWithSize } from './utils';
-
-interface Props {
-  items: IVideoElement[];
-}
-
-export default function InlinedVideoList({ items }: Props) {
-  return (
-    <ListWithSize<IVideoElement>
-      elements={items}
-      renderElement={VideoElement}
-    />
-  );
-}
